Fixing LogOut
