+++ app/src/main/java/edu/eci/ieti/takeiteasysk/ui/products/ProductsActivity.java
@@ -1,6 +1,8 @@
 package edu.eci.ieti.takeiteasysk.ui.products;
 
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -11,6 +13,7 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import edu.eci.ieti.takeiteasysk.MainActivity;
 import edu.eci.ieti.takeiteasysk.NewProductForm;
 import edu.eci.ieti.takeiteasysk.R;
 import edu.eci.ieti.takeiteasysk.persistence.ProductsPersistence;
@@ -52,6 +55,13 @@ public class ProductsActivity extends AppCompatActivity {
         LinearLayoutManager layoutManager = new LinearLayoutManager( this );
         recyclerView.setAdapter( productsAdapter );
         recyclerView.setLayoutManager(layoutManager);
+    }
+    public void onLogoutClicked(View view){
+        SharedPreferences sharedPref = getSharedPreferences(getString(R.string.preferece_file_key), Context.MODE_PRIVATE );
+        sharedPref.edit().clear().commit();
+        Intent intent = new Intent(this, MainActivity.class);
+        startActivity(intent);
+        finish();
     }
 
 }
